[RFC] nssearch: Improve performance of AcpiNsSearchOneScope
As documented, the linear search performed by AcpiNsSearchOneScope is
good enough for small namespace trees and is generally dominated by
other overheads anyway.

However, there are environments where improved performance during boot
is highly desirable.

This patch is a proposed path forward towards improving performance by
using a hash table to look up nodes. The hash table is keyed by a
combination of node name and parent node pointer.

Instead of a fixed number of hash table buckets, it might be preferable
to pick the size dynamically: for example, allocation can be shifted
until after the total size of ACPI tables is known, and that can be used
for a heuristic to pick the size and avoid extraneous memory overhead.

The new functionality is guarded behind a macro,
ACPI_USE_NS_SEARCH_ACCELERATION, which should be set in the
platform-specific headers along with the various dependencies
(exemplified here for the Linux kernel case in aclinux.h).

Signed-off-by: Peter Szuecs <pszucs@amazon.de>
Signed-off-by: Stanislav Spassov <stanspas@amazon.de>