tests/common: use CopyObject instead of Store for NRMT
NRMT is initialized as "", which is a zero-length string. Trying to
Store a longer string to it works on ACPICA, however, is incompatible
with other implementations like Acpi.sys, which doesn't allow resizing
existing buffers with Store. Therefore any Store operation to it
produces the same result on those implementations.

Use CopyObject to avoid any possible behavior differences.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>