MPAM: Fix table offsets
By convention, the various ACPI table dumps prints the table offset
from the top of the table. On the other hand, the MPAM table dump
prints sub-table offsets which make it harder to locate the
corresponding entry in the binary table.

Fix this by updating the AcpiDmDumpMpam() function to take into
account offsets from the top of the table when decoding sub-table
values.

Signed-off-by: Punit Agrawal <punit.agrawal@bytedance.com>