Anonymous ResourceTemplate dissassembly fix
Previous fix at 09a9a0808ba4997173ecb5c6ec3bdb1f2d0d50b5 was intended to capture the AcpiDmPredefinedDescription call in the conditional but the original call was left in place. This code still results in ASAN reporting a heap overflow on the minimal test case unless the original call is removed.