acpica: Revert "Do not cross page boundaries when mapping operation regions."
Revert changes made in commit 381f8561ee4e ("Do not cross page
boundaries when mapping operation regions."). The original intention
of that commit was to prevent memory mappings for operation regions from
crossing page boundaries, as it can cause warnings in case of different
page attributes.

However, it was discovered that when this condition occurs, mapping is
done until the end of the boundary, but still attempts to read/write
the entire length of the map, resulting in a NULL pointer deference. For
instance, if there is a request to map four bytes but only one byte is
mapped because it reaches the end of the current page boundary, there is
still an attempt to read/write four bytes, causing a NULL pointer
dereference.

Instead, map the complete length, as there is no restriction in the ACPI
specification that requires it to be in the same page boundary. It is
acceptable for it to be mapped in different regions.

Cc: Mario Limonciello <mario.limonciello@amd.com>
Link: https://bugzilla.kernel.org/show_bug.cgi?id=218849
Link: https://bugzilla.kernel.org/show_bug.cgi?id=14445
Fixes: 381f8561ee4e ("Do not cross page boundaries when mapping operation regions.")
Co-developed-by: Sanath S <Sanath.S@amd.com>
Signed-off-by: Sanath S <Sanath.S@amd.com>
Signed-off-by: Raju Rangoju <Raju.Rangoju@amd.com>