PHAT FW health table can be zero-length
When calculating the VendorLength of the PHAT FW health data subtable,
the result becomes a negative integer.  However, since UINT32 is being
used, it looks like a huge positive integer instead.  Conditionalize
the length calculation to handle this case properly.

This was discovered by using the command: iasl -T phat

Signed-off-by: Al Stone <ahs3@ahs3.net>