PHAT: Prevent segfault when encountering invalid device paths
If the device path inside a firmware health record is not properly
NULL-terminated, wcslen() will potentially return an invalid path
length, which can cause a segfault later when dumping the path data.

Fix this by using the device-specific data offset to calculate the
path length. This also allows us to warn if the device-specific data
offset is invalid.

Signed-off-by: Armin Wolf <W_Armin@gmx.de>