iASL/Templates: Fix CDAT template generation with DtCreateAllTemplates
When generating all templates with <-T ALL>, CDAT fails with an invalid
signature error message. This is due to the fact that CDAT does not have
an actual signature field in the table header, and its signature
contains at least one non-ASCII character that throws the error. To fix
this, compare TableData->Signature with "CDAT", and handle it as a
special table by dumping with AcpiDmDumpCdat instead of defaulting to
AcpiDmDumpDataTable.

Signed-off-by: Ahmed Salem <x0rw3ll@gmail.com>