Switch back to using strncpy in acpi_ut_safe_strncpy
I mistakenly replaced strncpy with memcpy in commit # 83019b4, not
realizing the entire context behind *why* strncpy was used. In this
safer implementation of strncpy, it does not make sense to use memcpy
only to null-terminate strings passed to acpi_ut_safe_strncpy one byte
early. The consequences of doing so are understandably *bad*, as was
evident by the kernel test bot reporting problems[1]

Link: https://lore.kernel.org/all/202505081033.50e45ff4-lkp@intel.com
[1]
Reported-by: kernel test robot <oliver.sang@intel.com>
Closes: https://lore.kernel.org/oe-lkp/202505081033.50e45ff4-lkp@intel.com

Signed-off-by: Ahmed Salem <x0rw3ll@gmail.com>