dispatcher: Use AcpiDsClearOperands() in AcpiDsCallControlMethod()
When deleting the previous walkstate operand stack
AcpiDsCallControlMethod() was deleting ObjDesc->Method.ParamCount
operands. But Method.ParamCount does not necessarily match
ThisWalkState->NumOperands, it may be either less or more.

After correcting the for loop to check `i < ThisWalkState->NumOperands`
the code is identical to AcpiDsClearOperands(), so just outright
replace the code with AcpiDsClearOperands() to fix this.

Signed-off-by: Hans de Goede <hansg@kernel.org>