ACPI 6.4: PPTT: include all fields in subtable type 1
In PPTT version 3 an extra field, Cache ID, was added to the Cache Type
Structure.  The struct, ACPI_PPTT_CACHE_V1, contains only this field. This
differs from the treatment of other versioned structures and is unexpected
for linux which reuses the actbl2.h header file. Include all the fields of
the new Cache Type Structure in ACPI_PPTT_CACHE_V1 and fix up all uses.

Signed-off-by: Ben Horgan <ben.horgan@arm.com>