Fix asltests using the Fatal() opcode
Some asltests test the behavior of the Fatal() opcode and thus require
that said opcode does not return an error when called.

Introduce a compile-time option called ACPI_CONTINUE_ON_FATAL to
instruct the executor to continue the execution of AML bytecode when
encountering a Fatal() opcode. Also update the asltest to use this
new option.

Fixes: d516c7758ba6 ("Abort AML bytecode execution when executing AML_FATAL_OP")
Signed-off-by: Armin Wolf <W_Armin@gmx.de>